+++ CHANGELOG.md
@@ -11,6 +11,16 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v4.0.0-alpha.?? (2024-??-??)
+
+#### :boom: Breaking Change
+
+* `core/lazy` no longer has a default export `core/lazy`
+
+#### :bug: Bug Fix
+
+* Added the `disposeLazy` function which clears the `makeLazy` module of the passed context to avoid memory leaks `core/lazy`
+
 ## v4.0.0-alpha.19 (2024-01-12)
 
 #### :bug: Bug Fix